feature: pull monaco API out and make it better
- replace `monaco.languages.graphql` with an `api` export object
- combine api and defaults into one interface
- get rid of `.d.ts` typings and module extension which caused issues for normal .ts typings
- update documentation to reflect API changes
