+++ examples/monaco-graphql-webpack/src/index.ts
@@ -1,12 +1,10 @@
 /* global monaco */
 
 import 'regenerator-runtime/runtime';
-import 'monaco-graphql/esm/monaco.contribution';
+import { api as GraphQLAPI } from 'monaco-graphql/esm/monaco.contribution';
 
 // eslint-disable-next-line spaced-comment
 /// <reference types='monaco-editor'/>
-// eslint-disable-next-line spaced-comment
-///  <reference types='monaco-graphql'/>
 
 // NOTE: using loader syntax becuase Yaml worker imports editor.worker directly and that
 // import shouldn't go through loader syntax.
@@ -42,7 +40,7 @@ schemaInput.onkeyup = e => {
   // @ts-ignore
   const val = e?.target?.value as string;
   if (val && typeof val === 'string') {
-    monaco.languages.graphql.graphqlDefaults.setSchemaConfig({ uri: val });
+    GraphQLAPI.setSchemaConfig({ uri: val });
   }
 };
 
@@ -99,15 +97,13 @@ const operationEditor = monaco.editor.create(
   },
 );
 
-monaco.languages.graphql.graphqlDefaults.setFormattingOptions({
-  // TODO: get these types working properly
+GraphQLAPI.setFormattingOptions({
   prettierConfig: {
-    printWith: 120,
+    printWidth: 120,
   },
 });
 
-monaco.languages.graphql.graphqlDefaults.setSchemaConfig({
-  // TODO: get these types working properly
+GraphQLAPI.setSchemaConfig({
   uri: SCHEMA_URL,
 });
 
@@ -124,7 +120,7 @@ async function executeCurrentOp() {
     if (parsedVariables && Object.keys(parsedVariables).length) {
       body.variables = variables;
     }
-    const result = await fetch(SCHEMA_URL, {
+    const result = await fetch(GraphQLAPI.schemaConfig.uri || SCHEMA_URL, {
       method: 'POST',
       headers: { 'content-type': 'application/json' },
       body: JSON.stringify(body),
