+++ packages/monaco-graphql/README.md
@@ -27,7 +27,7 @@ yarn add monaco-graphql
 ```ts
 import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
 
-import 'monaco-graphql/esm/monaco.contribution';
+import { api as GraphQLAPI } from 'monaco-graphql/esm/monaco.contribution';
 
 import EditorWorker from 'worker-loader!monaco-editor/esm/vs/editor/editor.worker';
 import GraphQLWorker from 'worker-loader!monaco-graphql/esm/graphql.worker';
@@ -45,16 +45,14 @@ monaco.editor.create(document.getElementById('someElementId'), {
   language: 'graphqlDev',
 });
 
-monaco.languages.graphql.graphqlDefaults.setSchemaUri(
-  'https://localhost:1234/graphql',
-);
+GraphQLAPI.setSchemaUri('https://localhost:1234/graphql');
 ```
 
 This will cover the basics, making an HTTP POST with the default `introspectionQuery()` operation. To customize the entire fetcher, see [advanced customization]() below
 
 ## Advanced Usage
 
-### `monaco.languages.graphql.graphqlDefaults` ([typedoc](http://graphiql-test.netlify/typedoc/interfaces/monaco_graphql.monaco.languages.graphql.languageservicedefaults))
+### `GraphQLAPI` ([typedoc](http://graphiql-test.netlify/typedoc/interfaces/monaco_graphql.api.languageserviceapi))
 
 If you call any of these API methods to modify the language service configuration at any point at runtime, the webworker will reload relevant language features.
 
@@ -65,23 +63,23 @@ verbs prefixes for methods are meaningful:
 - `set...` means to force re-write the whole settings entry for that method
 - `update...` means a shallow merge of the object/value you pass with the rest of the existing settings
 
-#### `graphqlDefaults.updateSchemaConfig()`
+#### `GraphQLAPI.updateSchemaConfig()`
 
 set schema `uri` (required) as well as `requestOptions`, `buildSchemaConfig` and `introspectionOptions`, with a shallow merge.
 invoking these will cause the webworker to reload language services
 
 ```ts
-monaco.languages.graphql.graphqlDefaults.updateSchemaConfig({
+GraphQLAPI.updateSchemaConfig({
   uri: '',
 });
 ```
 
-#### `graphqlDefaults.setSchemaConfig()`
+#### `GraphQLAPI.setSchemaConfig()`
 
 same as the above, except it overwrites the entire schema config
 
 ```ts
-monaco.languages.graphql.graphqlDefaults.updateSchemaConfig({
+GraphQLAPI.updateSchemaConfig({
   uri: 'https://my/schema',
   requestOptions: {
     headers: { Authorization: 'Bear Auth 2134' },
@@ -89,22 +87,20 @@ monaco.languages.graphql.graphqlDefaults.updateSchemaConfig({
 });
 ```
 
-#### `graphqlDefaults.setSchemaUri()`
+#### `GraphQLAPI.setSchemaUri()`
 
 You can also just change the schema uri directly!
 
 ```ts
-monaco.languages.graphql.graphqlDefaults.setSchemaUri(
-  'https://localhost:1234/graphql',
-);
+GraphQLAPI.setSchemaUri('https://localhost:1234/graphql');
 ```
 
-#### `graphqlDefaults.setModeConfiguration()`
+#### `GraphQLAPI.setModeConfiguration()`
 
 This is where you can toggle monaco language features. all are enabled by default.
 
 ```ts
-monaco.languages.graphql.graphqlDefaults.setModeConfiguration({
+GraphQLAPI.setModeConfiguration({
   documentFormattingEdits: true;
   completionItems: true;
   hovers: true;
@@ -113,7 +109,7 @@ monaco.languages.graphql.graphqlDefaults.setModeConfiguration({
 })
 ```
 
-#### `graphqlDefaults.setFormattingOptions()`
+#### `GraphQLAPI.setFormattingOptions()`
 
 this accepts an object `{ prettierConfig: prettier.Options }`, which accepts [any prettier option](https://prettier.io/docs/en/options.html).
 it will not re-load the schema or language features, however the new prettier options will take effect.
@@ -121,19 +117,17 @@ it will not re-load the schema or language features, however the new prettier op
 this method overwrites the previous configuration, and will only accept static values that can be passed between the main/worker process boundary.
 
 ```ts
-graphqlDefaults.setFormattingOptions({
+GraphQLAPI.setFormattingOptions({
   // if you wanna be like that
   prettierOptions: { tabWidth: 2, useTabs: true },
 });
 ```
 
-### `monaco.languages.graphql.api` ([typedoc](http://graphiql-test.netlify/typedoc/classes/monaco_graphql.monacographqlapi))
-
-#### `api.getSchema()`
+#### `GraphQLAPI.getSchema()`
 
 Returns either an AST `DocumentNode` or `IntrospectionQuery` result json using default or provided `schemaLoader`
 
-### `api.parse()`
+### `GraphQLAPI.parse()`
 
 parse graphql from string using webworkers (less render-blocking/multi-threaded CPU/etc)
